tested by faked corr

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: kclaudan <kclaudan@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/15 16:51:06 by kclaudan          #+#    #+#             */
-/*   Updated: 2024/11/15 16:51:49 by kclaudan         ###   ########.fr       */
+/*   Updated: 2024/11/16 11:43:02 by kclaudan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@
 void	ft_putchar(const char c);
 int		ft_putstr(const char *str);
 int		format_print(char f, va_list *args);
-int		ft_putnbr_base(unsigned int nbr, int r, int maj);
+int		ft_putnbr_base(unsigned long nbr, int r, int maj);
 int		print_u(unsigned int n);
 int		print_d(int n, int fd, int counter);
 int		is_format_specifier(const char f);
